Refactor orbit camera to use double precision and improve panning logic.
Introduces a new OrbitCameraPan gesture system to prevent drift during interaction and updates geometry types (Rect2, Bounds2) to support double precision. The change also migrates core camera calculations from float to double for improved numerical stability.

+++ graphics/tcplot/include/tcplot/orbit_camera.hpp
@@ -7,8 +7,8 @@ namespace tcplot {
 
     class TCPLOT_API OrbitCamera : public termin::OrbitCamera {
     public:
-        float& near;
-        float& far;
+        double& near;
+        double& far;
 
         OrbitCamera()
             : termin::OrbitCamera(),
